+++ docs/scripts/config/lang.js
@@ -1,0 +1,19 @@
+import en from '../lang/en.json';
+import cn from '../lang/cn.json';
+
+// for link
+const lang = {
+  en: 'English',
+  cn: '中文'
+};
+
+// ready translated locales
+const locales = {
+  en,
+  cn
+};
+
+export {
+  lang,
+  locales
+};