+++ docs/scripts/routes/layout.js
@@ -1,0 +1,20 @@
+// const Layout = require('../views/layout');
+const Grid = require('../views/grid');
+const Tabs = require('../views/tabs');
+const Footer = require('../views/footer');
+
+const layoutRoutes = [{
+  path: '/grid',
+  name: 'grid',
+  component: Grid
+}, {
+  path: '/tabs',
+  name: 'tabs',
+  component: Tabs
+}, {
+  path: '/footer',
+  name: 'footer',
+  component: Footer
+}];
+
+export default layoutRoutes;