+++ src/components/table.vue
@@ -20,14 +20,14 @@
               v-if="cell.isCheckbox"
               :value="cell.value"
               :model="isCheckAll"
-              @input="onCheckAll"></ui-checkbox>
+              @change="onCheckAll"></ui-checkbox>
           </th>
         </tr>
       </thead>
     </slot>
     <slot name="tbody" :data="tbodyData">
       <tbody>
-        <tr v-for="row in tbodyData">
+        <tr v-if="tbodyData.length" v-for="row in tbodyData">
           <td v-for="cell in row"
             :colspan="cell.col"
             :rowspan="cell.row"
@@ -36,15 +36,19 @@
             <ui-checkbox name="checkOne[]"
               v-if="cell.isCheckbox"
               :value="cell.value"
-              :model="currentCheckList"
-              @input="onCheckOne"></ui-checkbox>
+              :model="currentCheckboxList"
+              @change="onCheckOne"></ui-checkbox>
             <ui-button v-if="cell.isAction"
               v-for="action in cell.actions"
-              v-html="action.value"
-              :icon="action.icon"
-              :link="action.link"
-              @click.native="doAction(action.name, action.data)"></ui-button>
+              :icon="action.icon || action.isIcon"
+              :link="action.isLink"
+              @click.native="doAction(action.name, action.data)">
+                <span v-if="!action.icon" v-html="action.value"></span>
+              </ui-button>
           </td>
+        </tr>
+        <tr v-if="!tbodyData.length">
+          <td class="mdl-no-data-table" :colspan="col">{{ noData }}</td>
         </tr>
       </tbody>
     </slot>
@@ -70,16 +74,19 @@ import UiCheckbox from './form/checkbox';
 const T_HEAD = 'thead';
 const T_BODY = 'tbody';
 const T_FOOT = 'tfoot';
-const CELL_VALUE = 'value';
+const CELL_CLASS = 'class';
 const CELL_COLSPAN = 'col';
+const CELL_DATA = 'data';
+const CELL_FIELD = 'field';
+const CELL_FUNCTION = 'fn';
+const CELL_ICON = 'icon';
+const CELL_INDEX = 'index';
 const CELL_ROWSPAN = 'row';
 const CELL_SORT = 'sort';
-const CELL_CLASS = 'class';
-const CELL_INDEX = 'index';
-const CELL_DATA = 'data';
-const ACTION_LINK = 'link';
+const CELL_VALUE = 'value';
 const ACTION_BUTTON = 'button';
 const ACTION_ICON = 'icon';
+const ACTION_LINK = 'link';
 const CHECKBOX_POSITION_LEFT = 'left';
 const CHECKBOX_POSITION_RIGHT = 'right';
 const AGGREGATE_COUNT = 'count';
@@ -90,6 +97,9 @@ const AGGREGATE_MAX = 'max';
 const SORT_ASC = 'asc';
 const SORT_DESC = 'desc';
 const SORT_BY = 'by';
+const CLASSNAME_NON_NUMERIC = 'mdl-data-table__cell--non-numeric';
+const CLASSNAME_TEXT_LEFT = 'mdl-data-table__cell--text-left';
+const CLASSNAME_TEXT_CENTER = 'mdl-data-table__cell--text-center';
 const CALLBACK_SELECTED = 'selected';
 
 /**
@@ -127,22 +137,31 @@ export default {
       type: [Array, Boolean],
       default: false
     },
-    action: Array,
-    selectable: {
-      type: [String, Boolean],
-      default: false
-    },
-    checkList: {
+    action: {
       type: Array,
       default: function() {
         return [];
       }
+    },
+    selectable: {
+      type: [String, Boolean],
+      default: false
+    },
+    checkboxList: {
+      type: Array,
+      default: function() {
+        return [];
+      }
+    },
+    noData: {
+      type: String,
+      default: 'No Data123'
     }
   },
   data() {
     return {
       isCheckAll: false,
-      currentCheckList: this.checkList,
+      currentCheckboxList: this.checkboxList,
       currentData: this.data,
       currentThead: this.thead
     };
@@ -151,49 +170,56 @@ export default {
     theadData() {
       return this.getData({
         type: T_HEAD,
-        table: this.currentThead,
-        data: this.currentData,
-        selectable: this.selectable
+        table: this.currentThead
       });
     },
     tbodyData() {
       return this.getData({
         type: T_BODY,
         table: this.tbody,
-        data: this.currentData,
-        selectable: this.selectable,
         action: this.action
       });
     },
     tfootData() {
-      return this.tfoot ? this.getData({
+      return this.tfoot && this.currentData.length ? this.getData({
         type: T_FOOT,
-        table: this.tfoot,
-        data: this.currentData,
-        selectable: this.selectable,
+        table: this.tfoot
       }) : [];
     }
   },
   methods: {
     getCell(type, data) {
       let cell = {};
+      let fn;
 
       if (isString(data)) {
         cell[CELL_VALUE] = data;
       } else if (isObject(data)) {
+        // colspan attribute
         if (data[CELL_COLSPAN]) {
           cell[CELL_COLSPAN] = data[CELL_COLSPAN];
         }
+        // rowspan attribute
         if (data[CELL_ROWSPAN]) {
           cell[CELL_ROWSPAN] = data[CELL_ROWSPAN];
         }
-
+        // class attribute
         let className = [];
         if (data[CELL_CLASS]) {
           className.push(data[CELL_CLASS]);
         }
         if (data.noNum) {
-          className.push('mdl-data-table__cell--non-numeric');
+          className.push(CLASSNAME_NON_NUMERIC);
+        }
+        if (data.align) {
+          switch (data.align.toLowerCase()) {
+            case 'left':
+              className.push(CLASSNAME_TEXT_LEFT);
+              break;
+            case 'center':
+              className.push(CLASSNAME_TEXT_CENTER);
+              break;
+          }
         }
         if (className.length) {
           cell[CELL_CLASS] = className.join(' ');
@@ -201,8 +227,10 @@ export default {
 
         switch (type) {
           case T_HEAD:
+            // index: value
             cell[CELL_INDEX] = data[CELL_INDEX];
             cell[CELL_VALUE] = data[CELL_VALUE];
+            // sort by
             if (data[CELL_SORT] && data[SORT_BY]) {
               cell[CELL_SORT] = data[CELL_SORT];
               cell[SORT_BY] = data[SORT_BY];
@@ -210,46 +238,50 @@ export default {
             break;
           case T_FOOT:
             let result = 0;
-            switch (data.name.toLowerCase()) {
-              case AGGREGATE_COUNT:
-                result = data[CELL_DATA].length;
-                break;
-              case AGGREGATE_SUM:
-                data[CELL_DATA].forEach(value => {
-                  result += value;
-                });
-                break;
-              case AGGREGATE_AVG:
-                data[CELL_DATA].forEach(value => {
-                  result += value;
-                });
-                result /= data[CELL_DATA].length;
-                break;
-              case AGGREGATE_MIN:
-                data[CELL_DATA].forEach(value => {
-                  if (value < result) {
-                    result = value;
-                  } else if (result === 0) {
-                    result = value;
-                  }
-                });
-                break;
-              case AGGREGATE_MAX:
-                 data[CELL_DATA].forEach(value => {
-                  if (value > result) {
-                    result = value;
-                  }
-                });
-                break;
-              default:
-                result = '';
-                break;
+            if (data[CELL_DATA].length) {
+              switch (data.name.toLowerCase()) {
+                case AGGREGATE_COUNT:
+                  result = data[CELL_DATA].length;
+                  break;
+                case AGGREGATE_SUM:
+                  data[CELL_DATA].forEach(value => {
+                    result += value;
+                  });
+                  break;
+                case AGGREGATE_AVG:
+                  data[CELL_DATA].forEach(value => {
+                    result += value;
+                  });
+                  result /= data[CELL_DATA].length;
+                  break;
+                case AGGREGATE_MIN:
+                  data[CELL_DATA].forEach(value => {
+                    if (value < result) {
+                      result = value;
+                    } else if (result === 0) {
+                      result = value;
+                    }
+                  });
+                  break;
+                case AGGREGATE_MAX:
+                   data[CELL_DATA].forEach(value => {
+                    if (value > result) {
+                      result = value;
+                    }
+                  });
+                  break;
+                default:
+                  result = '';
+                  break;
+              }
             }
-            // TODO: format
-            cell[CELL_VALUE] = result;
+            fn = data[CELL_FUNCTION];
+            cell[CELL_VALUE] = fn ? fn(result) : Math.round(result * 100) / 100;
             break;
-          default:
-            cell[CELL_VALUE] = data[CELL_VALUE];
+          default: // T_BODY
+            fn = data[CELL_FUNCTION];
+            let val = data[CELL_VALUE];
+            cell[CELL_VALUE] = fn ? fn(val) : val;
             break;
         }
       } else {
@@ -260,28 +292,30 @@ export default {
       return cell;
     },
     getCheckbox(type, result, key = 1) {
-      let cell = {};
+      if (this.currentData.length) {
+        let cell = {};
 
-      switch (type) {
-        case T_HEAD:
-          cell = {
-            row: key, // row number
-            isCheckbox: true,
-            value: true
-          };
-          break;
-        case T_BODY:
-          cell = {
-            isCheckbox: true,
-            value: key // data[this.keyField]
-          };
-          break;
-      }
+        switch (type) {
+          case T_HEAD:
+            cell = {
+              row: key, // row number
+              isCheckbox: true,
+              value: true
+            };
+            break;
+          case T_BODY:
+            cell = {
+              isCheckbox: true,
+              value: key // data[this.keyField]
+            };
+            break;
+        }
 
-      if (this.selectable === CHECKBOX_POSITION_LEFT) {
-        result.unshift(cell);
-      } else if (this.selectable === CHECKBOX_POSITION_RIGHT) {
-        result.push(cell);
+        if (this.selectable === CHECKBOX_POSITION_RIGHT) {
+          result.push(cell);
+        } else if (this.selectable === CHECKBOX_POSITION_LEFT || this.selectable) {
+          result.unshift(cell);
+        }
       }
 
       return result;
@@ -291,15 +325,16 @@ export default {
       for (let action of this.action) {
         let cellData = {
           name: action.name,
-          value: action[CELL_VALUE],
+          value: action[CELL_VALUE] || action.name,
           data: data
         };
         switch (action.type.toLowerCase()) {
           case ACTION_LINK:
-            cellData.link = true;
+            cellData.isLink = true;
             break;
           case ACTION_ICON:
-            cellData.icon = true;
+            cellData.isIcon = true;
+            cellData[CELL_ICON] = action[CELL_ICON];
             break;
         }
         actions.push(cellData);
@@ -319,7 +354,6 @@ export default {
 
       if (Array.isArray(table)) {
         let cell;
-        let isMultiLine = false;
 
         switch (type) {
           case T_HEAD:
@@ -327,7 +361,6 @@ export default {
             for (let index in table) {
               let row = table[index];
               if (Array.isArray(row)) { // multi line
-                isMultiLine = true;
                 result[index] = [];
                 for (let key in row) {
                   let value = row[key];
@@ -341,49 +374,63 @@ export default {
                   result[index].push(this.getCell(type, cell));
                 }
               } else { // single line
+                let value = row;
                 cell = {};
-                if (isObject(row)) {
-                  cell = row;
+                if (isObject(value)) {
+                  cell = value;
                 } else {
-                  cell[CELL_VALUE] = row;
+                  cell[CELL_VALUE] = value;
                 }
                 cell[CELL_INDEX] = index;
                 result[0].push(this.getCell(type, cell));
               }
             }
             // add checkbox
-            if (object.selectable) {
+            if (this.selectable) {
               result[0] = this.getCheckbox(type, result[0], table.length);
             }
             break;
           case T_BODY:
-            for (let key in object.data) {
-              let value = object.data[key];
+            for (let key in this.currentData) {
+              let value = this.currentData[key];
               result[key] = [];
-              for (let field of table) {
-                cell = {
-                  value: value[field] // set value
-                };
+              // fill for cell
+              for (let item of table) {
+                cell = {};
+                if (isObject(item)) {
+                  cell = item;
+                  cell[CELL_VALUE] = value[item[CELL_FIELD]]; // use `field` !important
+                } else {
+                  cell[CELL_VALUE] = value[item];
+                }
                 result[key].push(this.getCell(type, cell));
               }
               // add action
               result[key] = this.getAction(result[key], value);
               // add checkbox
-              result[key] = this.getCheckbox(type, result[key], value[this.keyField]);
+              if (this.selectable) {
+                result[key] = this.getCheckbox(type, result[key], value[this.keyField]);
+              }
             }
             break;
           case T_FOOT:
             for (let item of table) {
-              cell = item || '';
-              let field = item ? item[CELL_VALUE] : '';
-              if (field) {
-                cell[CELL_DATA] = item ? this.currentData.map(value => value[field]) : '';
+              if (item) {
+                cell = item;
+                let field = item && item[CELL_FIELD] ? item[CELL_FIELD] : false;
+                if (field) {
+                  cell[CELL_DATA] = item ? this.currentData.map(value => value[field]) : '';
+                }
+                result.push(this.getCell(type, cell));
+              } else {
+                result.push({});
               }
-              result.push(this.getCell(type, cell));
             }
             // add empty action & checkbox
             result.push({});
-            result = this.getCheckbox(type, result);
+            if (this.selectable) {
+              result = this.getCheckbox(type, result);
+            }
             break;
         }
       } else {
@@ -396,16 +443,16 @@ export default {
       this.$emit(name, data);
     },
     onCheckOne(data) {
-      this.currentCheckList = data;
+      this.currentCheckboxList = data;
     },
     onCheckAll(checked) {
       this.isCheckAll = checked;
     },
     checkAll() {
       let dataCount = this.currentData.length; // not empty
-      let beEqual = this.currentCheckList.length === dataCount;
+      let beEqual = this.currentCheckboxList.length === dataCount;
       let ids = this.currentData.map(value => value[this.keyField]);
-      let exists = this.currentCheckList.every(id => ids.indexOf(id) > -1);
+      let exists = this.currentCheckboxList.every(id => ids.indexOf(id) > -1);
 
       this.isCheckAll = dataCount && beEqual && exists;
     },
@@ -436,16 +483,24 @@ export default {
     }
   },
   watch: {
-    checkList(val) {
-      this.currentCheckList = val;
+    data(val) {
+      this.currentData = val;
     },
-    currentCheckList(val) {
+    checkboxList(val) {
+      this.currentCheckboxList = val;
+    },
+    currentCheckboxList(val) {
       this.checkAll();
       this.$emit(CALLBACK_SELECTED, val);
     },
     isCheckAll(val) {
-      let lastCheckList = (this.currentCheckList.length === this.currentData.length) ? [] : this.currentCheckList;
-      this.currentCheckList = val ? this.currentData.map(value => value[this.keyField]) : lastCheckList;
+      let lastCheckList = (this.currentCheckboxList.length === this.currentData.length)
+        ? []
+        : this.currentCheckboxList;
+
+      this.currentCheckboxList = val
+        ? this.currentData.map(value => value[this.keyField])
+        : lastCheckList;
     },
     thead(val) {
       this.currentThead = val;
@@ -461,7 +516,5 @@ export default {
 </script>
 
 <style scoped>
-table {
-  width: 100%;
-}
+
 </style>